change representation of RGB values to use doubles in range 0.0-1.0

+++ src/RGB.hpp
@@ -2,11 +2,12 @@
 #define RGB_H
 
 // This structure represents a color with red, green, and blue components. 
+// Each component is represented by values between 0.0 and 1.0 inclusive.
 struct RGB {
-	int r;
-	int g; 
-	int b; 
-	RGB(int r = 0, int g = 0, int b = 0) : r(r), g(g), b(b) {}
+	double r;
+	double g; 
+	double b; 
+	RGB(double r = 0, double g = 0, double b = 0) : r(r), g(g), b(b) {}
 }; 
 
 #endif
